% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaeis_flowers.R
\name{elaeis_flowers}
\alias{elaeis_flowers}
\title{N\enc{ú}{u}mero de Flores de Cultivares de Dendezeiro ao
    Longo do Ano}
\format{Um \code{data.frame} com 611 observações e 8 variáveis, em
     que

\describe{

\item{\code{cult}}{Fator categórico que representa as cultivares de
    dendezeiro (\emph{Elaeis guineensis}).}

\item{\code{bloc}}{Fator categórico que representa os blocos do
    delineamento experimental.}

\item{\code{plant}}{Número inteiro que distingue entre as plantas da
    mesma parcela.}

\item{\code{days}}{Dias a partir de 10/01/2013, em que são feitas as
    avaliações nas plantas. As avaliação são em média a cada 30 dias
    e encerraram-se em 02/01/2014.}

\item{\code{tot}}{Número total de flores vistas na planta,
    independente do sexo da flor.}

\item{\code{female}}{Número de flores fêmea encontradas na
    planta. Valores \code{NA} representar que não foi possível
    distinguir o sexo das flores.}

\item{\code{male}}{Número de flores macho encontradas na
    planta. Valores \code{NA} representar que não foi possível
    distinguir o sexo das flores.}

\item{\code{abort}}{Número de flores abortadas vistas na
    planta. Valores \code{NA} indicam que essa variável não estava
    sendo observada.}

}}
\source{
Gustavo Azevedo Campos (\email{gustavo.campos@embrapa.br}).
}
\description{
Quatro cultivares de dendezeiro (\emph{Elaeis
    guineensis}) foram estudas em experimento em delineamento de
    blocos casualizados. Foi observado o número de flores fêmea e
    macho em avaliações mensais por um ano em três plantas em cada
    parcela.
}
\examples{

data(elaeis_flowers)
str(elaeis_flowers)

elaeis_flowers$ue <- with(elaeis_flowers,
                          interaction(cult, bloc, plant,
                                      drop = TRUE))
levels(elaeis_flowers$ue) <- 1:nlevels(elaeis_flowers$ue)
xtabs(~ue, data = elaeis_flowers)

library(lattice)

xyplot(I(male + female) ~ days | cult,
       groups = bloc,
       data = elaeis_flowers,
       type = c("p", "a"))

xyplot(male + female ~ days | ue,
       data = elaeis_flowers,
       type = "o",
       as.table = TRUE)

}

