% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_ake_b.R
\name{quality_ake_b}
\alias{quality_ake_b}
\title{Effect of fungicide sprays programs and pistachio hedging on
    fruit stain levels caused by \emph{Alternaria} late blight in
    commercial pistachio orchard of Tulare County, California.}
\format{Um \code{data.frame} com 288 observações e 12 variáveis, em
    que

\describe{

\item{\code{yr}}{Factor variable to indicate the year of experiment.}

\item{\code{hed}}{Factor variable to indicate the hedging type on
    trees.  The hedging is the shape the branchs and limbs when
    pruned. They can be heavy (severe pruning) or normal (regular
    pruning).}

\item{\code{tra}}{Factor variable to indicate the fungicide treatment
    on field, already described above. the treatment combines the use
    of one SDHI fungicide (Merivon or Fontelis) and one or two
    additional chemical group.}

\item{\code{plo}}{Not an important variable to consider in the
    analysis.  The plot code simply indicate the location of
    experimental plots, they are represented by one number (from 1 to
    3) indicating the rows (for example: each number is composed by 4
    rows from where the two central rows contain the flagged tree),
    and letters (from A to D) indicating the change in hedging.}

\item{\code{tre}}{Factor variable to indicate tree from where fruits
    were collected. There are 48 trees identified from 1 to 48, for
    2015, and the same identification for 2016 (corresponding to the
    same trees sampled one year later).  As an example: plot "1A",
    include trees 1, 2, 3 and 4 (1, 2, 3 treated and 4 not treated);
    and plot "1B" includes trees 5, 6, 7 and 8 (5, 6, 7 treated and 8
    not treated).}

\item{\code{rep}}{Integer variable to indicate repetition. Each fruit
    sample obtained from a certain tree was sub-sampled by randomly
    choosing 100 fruits.  This subsampling originated 3 repetitions
    for each tree called subsample 1, 2 and 3.}

\item{\code{c0}}{Numeric variable that is the number of fruits at
    stain category 0. Category zero are the number of pistachio
    fruits with 0\% of the shell surface discolored.}

\item{\code{c1}}{Numeric variable that is the number of fruits at
    stain category 1. Category one are the number of pistachio fruits
    with 1\% to 10\% of the shell surface discolored.}

\item{\code{c2}}{Numeric variable that is the number of fruits at
    stain category 2. Category two are the number of pistachio fruits
    with 11\% to 35\% of the shell surface discolored.}

\item{\code{c3}}{Numeric variable that is the number of fruits at
    stain category 3. Category three are the number of pistachio
    fruits with 36\% to 64\% of the shell surface discolored.}

\item{\code{c4}}{Numeric variable that is the number of fruits at
    stain category 4. Category four are the number of pistachio
    fruits with 65\% to 100\% of the shell surface discolored.}

\item{\code{tot}}{Numeric variable that is the total number of fruits
   evaluated inside each sub-sample.}

}}
\source{
Paulo dos Santos Faria Lichtemberg\eqn{^1}
    (\url{http://lattes.cnpq.br/8132272273348880}), Ryan D. Puckett
    (\url{http://kare.ucanr.edu/}), Walmes Marques Zeviani\eqn{^2}
    (\url{http://www.leg.ufpr.br/~walmes}), Connor G. Cunningham
    (\url{http://kare.ucanr.edu/}), Themis J. Michailides
    (\url{http://ucanr.edu/?facultyid=1535}).  \eqn{^1}University of
    California, Department of Plant Pathology, Kearney agricultural,
    research and extension center, 9240 S Riverbend Ave, Parlier,
    California, US.  \eqn{^2}Universidade Federal do Paraná,
    Departamento de Estatística.
}
\description{
The experiment was established in a commercial pistachio
    orchard in Tulare County, California. The total area sizes 2.9 ha
    with density of 335 plants/ha. In total, 12 plots were set with
    approximately 80 pistachio trees cv. Kerman spaced at 5.8 m
    between rows and 5.2 m between plants. Each plot included four
    rows in width.  In the year 2015, a heavy and normal hedging were
    intercalated, resulting in 6 plots for each hedging type.  No
    hedging was made in 2016.  Within each plot, a fungicide free
    sub-plot with 20 plants (4 rows by 5 plants) was set as a
    control.  Along the plot central rows, four plants (three inside
    the treated plot and one inside the control sub-plot) were
    identified for isolate collection, severity and defoliation
    assessments.  In total, 48 plants were identified. The fungicide
    treatment included three different programs; treatment 1: Merivon
    (1st application) and Switch (2nd application); treatment 2:
    FontelisTM (1st application) and Switch (2nd application); and
    treatment 3: Merivon (1st application), Switch (2nd application)
    and Gem (third application).  The combination of spray program
    and hedging type allowed two replication plots per treatment.  To
    determine the effect of spray program and pistachio hedging on
    fruit quality, each of the 48 flaged trees had their fruits
    harvested in mid September.  Fruits were bagged in plastic and
    identified to allow the localization of plot and tree. Mechanic
    dehulling was made prior to fruit drying, which occured during 3
    days at temperature of 55 \eqn{^\circ}C.  After drying, fruits
    were re-placed into an identified plastic bags and stored at cold
    room at 4 \eqn{^\circ}C until analyses. For each one of the 48
    bags of samples, three subsamples were prepared including 100
    fruits randomly choosen.  In order to evaluate the amount of
    stain caused by the incidence of \emph{Alternaria} late blight in
    the field, each subsample (composed by 100 fruits) were separated
    in 5 categories, from 0 (no stain) to 4 (more stain).
}
\examples{

data(quality_ake_b)
str(quality_ake_b)

library(reshape)

db <- melt(quality_ake_b[, -ncol(quality_ake_b)],
           id.vars = 1:6,
           id.measure = grep("c\\\\d", names(quality_ake_b)))
str(db)
levels(db$variable)

names(db)[ncol(db) - 1:0] <- c("categ", "freq")

library(lattice)
library(latticeExtra)

useOuterStrips(
    xyplot(freq ~ categ | hed + factor(yr),
           groups = tra,
           data = db,
           xlab = "Treatments",
           ylab = "Number of fruits",
           jitter.x = TRUE,
           auto.key = TRUE,
           type = c("p", "a")))

useOuterStrips(
    xyplot(freq ~ categ | tra + factor(yr),
           groups = hed,
           data = db,
           xlab = "Hed",
           ylab = "Number of fruits",
           jitter.x = TRUE,
           auto.key = TRUE,
           type = c("p", "a")))

}
