% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wzCoop.R
\docType{package}
\name{RDASC}
\alias{RDASC}
\alias{RDASC-package}
\title{Reproducible Data Analysis of Scientific Cooperations}
\description{
\code{RDASC} is a Reproducible Data Analysis
    Package. This is an R package to keep, in a reproducible and free
    accessible way, cientific colaborations that I
    participated. Datasets of each cooperation are documented and its
    analysis are done as package vignettes.
}
